Revert "fetch events and nosNotifications on notificationView"
This reverts commit afc211d6969c6357d34196b0aaba91d9e99cb4b1.
