+++ Nos/Models/CoreData/Event+CoreDataClass.swift
@@ -607,12 +607,4 @@ public class Event: NosManagedObject, VerifiableEvent {
         shouldBePublishedTo = Set()
     }
 }
-
-extension Event: NotificationDisplayable {
-    /// Returns self since an Event is its own associated event.
-    var event: Event? {
-        self
-    }
-}
-
 // swiftlint:enable file_length
